+++ app/src/main/java/pro/apir/tko/data/framework/network/model/request/RouteTrackingStartRequest.kt
@@ -1,3 +1,3 @@
 package pro.apir.tko.data.framework.network.model.request
 
-data class RouteTrackingStartRequest(val route: Int)
+data class RouteTrackingStartRequest(val route: Long)
